pull out generic gen funcs
