+++ src/boh_app/data/generate_items.py
@@ -1,21 +1,11 @@
 import json
-from enum import StrEnum
 from pathlib import Path
 from typing import Any
 
 from ..settings import CACHE_DIR
-from ..utils import find_boh_dir
+from .gen_utils import SteamFiles, get_steam_data, get_valid_refs
 
 HERE = Path(__file__).parent
-
-
-class SteamFiles(StrEnum):
-    ITEM = "aspecteditems.json"
-    INHERIT = "_prototypes.json"
-    SOUL1 = "abilities.json"
-    SOUL2 = "abilities2.json"
-    SOUL3 = "abilities3.json"
-    SOUL4 = "abilities4.json"
 
 
 def gen_items_json():
@@ -44,15 +34,6 @@ def get_soul_data():
     return soul_items
 
 
-def get_steam_data(selection: SteamFiles) -> list[dict[str, Any]]:
-    boh_data_dir = find_boh_dir() / "StreamingAssets/bhcontent/core/elements"
-    boh_file = boh_data_dir / selection.value
-
-    with boh_file.open() as a:
-        data = json.load(a)
-    return data["elements"]
-
-
 def prune_data(data: list[dict[str, Any]]):
     discarded = ["comfort", "bust", "cache", "spintria", "wallart"]
     beast_discarded = [".wild", ".hungry", "savage."]
@@ -66,12 +47,6 @@ def prune_data(data: list[dict[str, Any]]):
             continue
         pruned.append(item)
     return pruned
-
-
-def get_valid_refs(name: str):
-    from boh_app.data.load_data import get_data
-
-    return [d["id"] for d in get_data(name)]
 
 
 def get_our_items():
